backend/requirements.txt
fastapi
uvicorn
pydantic
openai
pydantic-settings
python-dotenv